Clear the clippy lint backlog outside the dead-code warnings
Mostly `cargo clippy --fix`: `map_or(true, ..)` to `is_none_or`, redundant
casts and conversions, `min().max()` to `clamp`, collapsed ifs, `is_multiple_of`,
a `Default` for `ThumbPool`, and `map` to `inspect` where the closure returned
its input untouched. Two by hand: Atlas's pre-warm walk took eight arguments,
so its four shared atomics now travel as `PrewarmTally`, and the deferred
canvas click vector gets a `DeferredClick` alias. No behaviour changes.

+++ apps/slate/src/app/board_path.rs
@@ -490,8 +490,8 @@ pub fn marquee_hits_node(node: &Node, marquee: WorldRect, zoom: f32) -> bool {
                 if let Some(bez) = bez_from_open_curve(node, s) {
                     let flat = flatten(&bez, 0.25);
                     for w in flat.windows(2) {
-                        let a = Pos2::new(w[0][0] as f32, w[0][1] as f32);
-                        let b = Pos2::new(w[1][0] as f32, w[1][1] as f32);
+                        let a = Pos2::new(w[0][0], w[0][1]);
+                        let b = Pos2::new(w[1][0], w[1][1]);
                         if segment_intersects_rect(a, b, marquee) {
                             return true;
                         }
